Refactor Note Services.
