+++ src/Service/Note/BaseNoteService.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace App\Service\Note;
 
+use App\Exception\Note;
 use App\Repository\NoteRepository;
 use App\Service\BaseService;
 use App\Service\RedisService;
+use Respect\Validation\Validator as v;
 
 abstract class BaseNoteService extends BaseService
 {
@@ -26,6 +28,29 @@ abstract class BaseNoteService extends BaseService
     {
         $this->noteRepository = $noteRepository;
         $this->redisService = $redisService;
+    }
+
+    protected static function validateNoteName(string $name): string
+    {
+        if (!v::length(2, 50)->validate($name)) {
+            throw new Note('The name of the note is invalid.', 400);
+        }
+
+        return $name;
+    }
+
+    public function getOneFromCache(int $noteId)
+    {
+        $redisKey = sprintf(self::REDIS_KEY, $noteId);
+        $key = $this->redisService->generateKey($redisKey);
+        if ($this->redisService->exists($key)) {
+            $note = $this->redisService->get($key);
+        } else {
+            $note = $this->getOneFromDb($noteId);
+            $this->redisService->setex($key, $note);
+        }
+
+        return $note;
     }
 
     public function getOneFromDb(int $noteId)
